+++ packages/components/src/components/carousel/carousel.tsx
@@ -2,7 +2,7 @@ import type { BreakpointCustomizable, PropTypes, Theme, ValidatorFunction } from
 import type { ButtonPure } from '../button-pure/button-pure';
 import type {
   CarouselAlignHeader,
-  CarouselChangeEvent,
+  CarouselUpdateEvent,
   CarouselInternationalization,
   CarouselWidth,
 } from './carousel-utils';
@@ -116,12 +116,12 @@ export class Carousel {
   @Prop() public activeSlideIndex?: number = 0;
 
   /**
-   * @deprecated since v3.0.0, will be removed with next major release, use `change` event instead.
+   * @deprecated since v3.0.0, will be removed with next major release, use `update` event instead.
    * Emitted when carousel's content slides. */
-  @Event({ bubbles: false }) public carouselChange: EventEmitter<CarouselChangeEvent>;
+  @Event({ bubbles: false }) public carouselChange: EventEmitter<CarouselUpdateEvent>;
 
   /** Emitted when carousel's content slides. */
-  @Event({ bubbles: false }) public change: EventEmitter<CarouselChangeEvent>;
+  @Event({ bubbles: false }) public update: EventEmitter<CarouselUpdateEvent>;
 
   @State() private amountOfPages: number;
 
@@ -288,7 +288,7 @@ export class Carousel {
       updatePrevNextButtons(this.btnPrev, this.btnNext, splide);
       updateSlidesInert(splide);
       updatePagination(this.paginationEl, activeIndex);
-      this.change.emit({ activeIndex, previousIndex });
+      this.update.emit({ activeIndex, previousIndex });
       this.carouselChange.emit({ activeIndex, previousIndex });
     });